move core chain logic to helper class ChainBase
